healthcare_project/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
python-decouple==3.8
requests==2.32.3